Closes #21. code review response
