+++ iis-core/src/test/java/eu/dnetlib/iis/core/OozieJobPropertiesFetcher.java
@@ -1,13 +1,8 @@
 package eu.dnetlib.iis.core;
 
-import java.io.File;
-import java.io.IOException;
 import java.io.StringReader;
 import java.util.Properties;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import org.apache.commons.io.FileUtils;
 import org.apache.oozie.client.OozieClient;
 import org.apache.oozie.client.WorkflowJob;
 import org.jdom.Document;
@@ -15,19 +10,24 @@ import org.jdom.Element;
 import org.jdom.input.SAXBuilder;
 
 /**
- * Class containing oozie related helper methods used for tests
+ * Fetcher of oozie job properties 
  * 
  * @author madryk
  *
  */
-class OozieTestHelper {
+class OozieJobPropertiesFetcher {
 
 	private OozieClient oozieClient;
 	
 	
-	public OozieTestHelper(OozieClient oozieClient) {
+	//------------------------ CONSTRUCTORS --------------------------
+	
+	public OozieJobPropertiesFetcher(OozieClient oozieClient) {
 		this.oozieClient = oozieClient;
 	}
+	
+	
+	//------------------------ LOGIC --------------------------
 	
 	/**
 	 * Returns properties of oozie job with provided id
@@ -56,25 +56,6 @@ class OozieTestHelper {
 		}
 		
 		return jobProperties;
-	}
-	
-	/**
-	 * Returns oozie job id based on log file
-	 */
-	public static String readJobIdFromLogFile(File runOozieJobLogFile) {
-		
-		String jobId;
-		try {
-			jobId = FileUtils.readFileToString(runOozieJobLogFile);
-		} catch (IOException e) {
-			throw new RuntimeException("Unable to read run oozie job log file", e);
-		}
-		Pattern pattern = Pattern.compile("^job: (\\S*)$", Pattern.MULTILINE);
-		Matcher matcher = pattern.matcher(jobId);
-		matcher.find();
-		jobId = matcher.group(1);
-		
-		return jobId;
 	}
 	
 }
